State: Move isDeactivatingJetpackModule to global selectors directory
